refactor: centralize plain_or_settings in config.py; prep rendering_tools imports
  - Move _plain_or_settings helper from economics/routines.py and
    infra/channel/email.py into apps/chat/sdk/config.py as single
    source of truth; both modules now import it from there
